Reorganize users views into separate version files
Including the tests
