+++ EquiTrack/users/tests/test_views.py
@@ -12,7 +12,6 @@ from EquiTrack.factories import CountryFactory, GroupFactory, OfficeFactory, Sec
 from EquiTrack.tests.mixins import APITenantTestCase, FastTenantTestCase
 from publics.tests.factories import BusinessAreaFactory
 from users.models import Group, User, UserProfile
-from users.serializers_v3 import AP_ALLOWED_COUNTRIES
 
 
 class TestUserAuthAPIView(APITenantTestCase):
@@ -20,7 +19,7 @@ class TestUserAuthAPIView(APITenantTestCase):
         self.user = UserFactory()
         response = self.forced_auth_req(
             "get",
-            "/users/api/profile/",
+            reverse("user-api-profile"),
             user=self.user
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
@@ -99,8 +98,7 @@ class TestOfficeViews(APITenantTestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
 
 
-class TestUserViews(APITenantTestCase):
-
+class TestStaffUsersView(APITenantTestCase):
     def setUp(self):
         self.unicef_staff = UserFactory(is_staff=True)
         self.unicef_superuser = UserFactory(is_superuser=True)
@@ -159,44 +157,6 @@ class TestUserViews(APITenantTestCase):
         response = self.forced_auth_req('get', '/api/groups/', user=self.unicef_staff)
 
         self.assertEqual(response.status_code, status.HTTP_200_OK)
-
-    def test_api_users_retrieve_myprofile(self):
-        response = self.forced_auth_req(
-            'get',
-            '/api/v2/users/myprofile/',
-            user=self.unicef_staff,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["name"], self.unicef_staff.get_full_name())
-
-    @skip('no update method on view')
-    def test_api_users_patch_myprofile(self):
-        data = {
-            "supervisor": self.unicef_superuser.id,
-            "oic": self.unicef_superuser.id,
-        }
-        response = self.forced_auth_req(
-            'patch',
-            '/api/v2/users/myprofile/',
-            user=self.unicef_staff,
-            data=data
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["supervisor"], self.unicef_superuser.id)
-        self.assertEqual(response.data["oic"], self.unicef_superuser.id)
-
-        response = self.forced_auth_req(
-            'get',
-            '/api/v2/users/myprofile/',
-            user=self.unicef_staff,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["user"], self.unicef_staff.id)
-        self.assertEqual(response.data["supervisor"], self.unicef_superuser.id)
-        self.assertEqual(response.data["oic"], self.unicef_superuser.id)
 
     def test_api_offices_detail(self):
         response = self.forced_auth_req('get', '/api/offices/', user=self.unicef_staff)
@@ -261,118 +221,6 @@ class TestUserViews(APITenantTestCase):
         response_json = json.loads(response.rendered_content)
 
         self.assertEqual(response_json['t2f']['business_area'], None)
-
-
-class TestUserViewsV3(APITenantTestCase):
-    def setUp(self):
-        self.unicef_user = UserFactory()
-        self.unicef_staff = UserFactory(is_staff=True)
-        self.unicef_superuser = UserFactory(is_superuser=True)
-        self.partnership_manager_user = UserFactory(is_staff=True)
-        self.group = GroupFactory()
-        self.partnership_manager_user.groups.add(self.group)
-
-    def test_api_users_list(self):
-        response = self.forced_auth_req('get', reverse('users_v3:users-list'), user=self.unicef_staff)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data), 2)
-
-    def test_api_users_list_user_forbidden(self):
-        response = self.forced_auth_req('get', reverse('users_v3:users-list'), user=self.unicef_user)
-        # non-staff users should not be able to see this endpoint
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_users_api_list_values(self):
-        response = self.forced_auth_req(
-            'get',
-            '/api/v3/users/',
-            user=self.unicef_staff,
-            data={"values": "{},{}".format(self.partnership_manager_user.id, self.unicef_superuser.id)}
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data), 1)
-
-    def test_api_users_list_values_bad(self):
-        response = self.forced_auth_req(
-            'get',
-            '/api/v3/users/',
-            user=self.unicef_staff,
-            data={"values": '1],2fg'}
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, [u'Query parameter values are not integers'])
-
-    def test_api_users_list_managers(self):
-        response = self.forced_auth_req(
-            'get',
-            '/api/v3/users/',
-            user=self.unicef_staff,
-            data={"partnership_managers": True}
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data), 2)
-
-    def test_api_users_retrieve_myprofile(self):
-        response = self.forced_auth_req(
-            'get',
-            '/api/v3/users/profile/',
-            user=self.unicef_staff,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["name"], self.unicef_staff.get_full_name())
-
-    def test_api_users_retrieve_myprofile_show_ap_false(self):
-        self.assertNotIn(self.unicef_staff.profile.country.name, AP_ALLOWED_COUNTRIES)
-        response = self.forced_auth_req(
-            'get',
-            reverse("users_v3:myprofile-detail"),
-            user=self.unicef_staff,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["show_ap"], False)
-
-    def test_api_users_retrieve_myprofile_show_ap(self):
-        self.unicef_staff.profile.country.name = AP_ALLOWED_COUNTRIES[0]
-        self.unicef_staff.profile.country.save()
-        self.assertIn(self.unicef_staff.profile.country.name, AP_ALLOWED_COUNTRIES)
-        response = self.forced_auth_req(
-            'get',
-            reverse("users_v3:myprofile-detail"),
-            user=self.unicef_staff,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["show_ap"], True)
-
-    def test_minimal_verbosity(self):
-        response = self.forced_auth_req('get', '/api/v3/users/',
-                                        data={'verbosity': 'minimal'}, user=self.unicef_superuser)
-
-        response_json = json.loads(response.rendered_content)
-        self.assertEqual(len(response_json), 2)
-
-    def test_retrieve_user_countries(self):
-        response = self.forced_auth_req('get', reverse('users_v3:country-detail'), user=self.unicef_user)
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data), 1)
-
-        self.assertEqual(self.unicef_user.profile.country.name, response.data[0]['name'])
-
-
-class TestCountryView(APITenantTestCase):
-    def test_get(self):
-        user = UserFactory(is_staff=True)
-        response = self.forced_auth_req(
-            "get",
-            reverse("users_v2:country-detail"),
-            user=user
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data[0]["name"], user.profile.country.name)
 
 
 class TestUsersDetailAPIView(APITenantTestCase):
