+++ EquiTrack/users/tests/test_views_v3.py
@@ -3,11 +3,14 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+import json
+
 from django.core.urlresolvers import reverse
 from rest_framework import status
 
-from EquiTrack.factories import UserFactory
+from EquiTrack.factories import GroupFactory, UserFactory
 from EquiTrack.tests.mixins import APITenantTestCase
+from users.serializers_v3 import AP_ALLOWED_COUNTRIES
 
 
 class TestCountryView(APITenantTestCase):
@@ -65,3 +68,95 @@ class TestUsersDetailAPIView(APITenantTestCase):
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         self.assertEqual(response.data, {})
+
+
+class TestUsersListAPIView(APITenantTestCase):
+    def setUp(self):
+        self.unicef_staff = UserFactory(is_staff=True)
+        self.unicef_superuser = UserFactory(is_superuser=True)
+        self.partnership_manager_user = UserFactory(is_staff=True)
+        self.group = GroupFactory()
+        self.partnership_manager_user.groups.add(self.group)
+        self.url = reverse("users_v3:users-list")
+
+    def test_api_users_list(self):
+        response = self.forced_auth_req('get', self.url, user=self.unicef_staff)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(len(response.data), 2)
+
+    def test_users_api_list_values(self):
+        response = self.forced_auth_req(
+            'get',
+            self.url,
+            user=self.unicef_staff,
+            data={"values": "{},{}".format(self.partnership_manager_user.id, self.unicef_superuser.id)}
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(len(response.data), 1)
+
+    def test_api_users_list_values_bad(self):
+        response = self.forced_auth_req(
+            'get',
+            self.url,
+            user=self.unicef_staff,
+            data={"values": '1],2fg'}
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertEqual(response.data, [u'Query parameter values are not integers'])
+
+    def test_api_users_list_managers(self):
+        response = self.forced_auth_req(
+            'get',
+            self.url,
+            user=self.unicef_staff,
+            data={"partnership_managers": True}
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(len(response.data), 2)
+
+    def test_api_users_retrieve_myprofile(self):
+        response = self.forced_auth_req(
+            'get',
+            reverse("users_v3:myprofile-detail"),
+            user=self.unicef_staff,
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["name"], self.unicef_staff.get_full_name())
+
+    def test_api_users_retrieve_myprofile_show_ap_false(self):
+        self.assertNotIn(self.unicef_staff.profile.country.name, AP_ALLOWED_COUNTRIES)
+        response = self.forced_auth_req(
+            'get',
+            reverse("users_v3:myprofile-detail"),
+            user=self.unicef_staff,
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["show_ap"], False)
+
+    def test_api_users_retrieve_myprofile_show_ap(self):
+        self.unicef_staff.profile.country.name = AP_ALLOWED_COUNTRIES[0]
+        self.unicef_staff.profile.country.save()
+        self.assertIn(self.unicef_staff.profile.country.name, AP_ALLOWED_COUNTRIES)
+        response = self.forced_auth_req(
+            'get',
+            reverse("users_v3:myprofile-detail"),
+            user=self.unicef_staff,
+        )
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        self.assertEqual(response.data["show_ap"], True)
+
+    def test_minimal_verbosity(self):
+        response = self.forced_auth_req(
+            'get',
+            self.url,
+            data={'verbosity': 'minimal'},
+            user=self.unicef_superuser
+        )
+        response_json = json.loads(response.rendered_content)
+        self.assertEqual(len(response_json), 2)
