+++ EquiTrack/users/urls_v3.py
@@ -1,7 +1,12 @@
 from django.conf.urls import url
-from views_v3 import CountryView, MyProfileAPIView, UsersDetailAPIView, UsersListApiView
 
-from users.views import ChangeUserCountryView
+from views_v3 import (
+    ChangeUserCountryView,
+    CountryView,
+    MyProfileAPIView,
+    UsersDetailAPIView,
+    UsersListAPIView,
+)
 
 urlpatterns = (
     # api
@@ -9,5 +14,5 @@ urlpatterns = (
     url(r'^changecountry/$', ChangeUserCountryView.as_view(http_method_names=['post']), name="country-change"),
     url(r'^country/$', CountryView.as_view(http_method_names=['get']), name="country-detail"),
     url(r'^(?P<pk>[0-9]+)/$', UsersDetailAPIView.as_view(http_method_names=['get']), name="user-detail"),
-    url(r'^$', UsersListApiView.as_view(), name="users-list"),
+    url(r'^$', UsersListAPIView.as_view(), name="users-list"),
 )
