+++ EquiTrack/users/views_v2.py
@@ -1,0 +1,49 @@
+from users import views as v1
+
+
+class UserAuthAPIView(v1.UserAuthAPIView):
+    """Stub for UserAuthAPIView"""
+
+
+class ChangeUserCountryView(v1.ChangeUserCountryView):
+    """Stub for ChangeUserCountryView"""
+
+
+class StaffUsersView(v1.StaffUsersView):
+    """Stub for UsersView"""
+
+
+class CountryView(v1.CountryView):
+    """Stub for CountryView"""
+
+
+class MyProfileAPIView(v1.MyProfileAPIView):
+    """stub for MyProfileAPIView"""
+
+
+class UsersDetailAPIView(v1.UsersDetailAPIView):
+    """Stub for UsersDetailAPIView"""
+
+
+class ProfileEdit(v1.ProfileEdit):
+    """Stub for ProfileEdit"""
+
+
+class GroupViewSet(v1.GroupViewSet):
+    """Stub for GroupViewSet"""
+
+
+class UserViewSet(v1.UserViewSet):
+    """Stub for UserViewSet"""
+
+
+class OfficeViewSet(v1.OfficeViewSet):
+    """Stub for OfficeViewSet"""
+
+
+class SectionViewSet(v1.SectionViewSet):
+    """Stub for SectionViewSet"""
+
+
+class ModuleRedirectView(v1.ModuleRedirectView):
+    """Stub for ModuleRedirectView"""
